requirements.txt
Django==5.0.6
django-debug-toolbar==4.4.6
Faker==12.0.1
mixer==7.2.2
psycopg2-binary==2.9.9
pylint==3.2.6
python-dotenv==1.0.1
coverage==7.6.0
redis==5.0.8
django-rq==2.10.2
djangorestframework==3.15.2
django-filter==24.3
markdown==3.7
djangorestframework-simplejwt==5.3.1
